feat(client): change directory timeline
