+++ client/src/app/(common)/_styles/common.css.ts
@@ -1,11 +1,15 @@
 import { style } from "@vanilla-extract/css";
 
-/* layout */
 export const containerStyle = style({
   paddingTop: "78px",
   paddingBottom: "56px",
   background: "#f9f8f4",
   minHeight: "100vh",
+});
+
+export const contentStyle = style({
+  margin: "24px 0",
+  padding: "0 24px",
 });
 
 export const navStyle = style({
@@ -24,22 +28,4 @@ export const navStyle = style({
 export const navOpenStyle = style({
   pointerEvents: "auto",
   opacity: 1,
-});
-
-/* page */
-export const contentStyle = style({
-  margin: "24px 0",
-  padding: "0 24px",
-});
-
-export const noticeStyle = style({
-  selectors: {
-    "&:not(:first-child)": {
-      marginTop: "24px",
-    },
-  },
-});
-
-export const labelStyle = style({
-  marginTop: "36px",
 });
