+++ client/src/app/(common)/timeline/_styles/timeline.css.ts
@@ -1,0 +1,13 @@
+import { style } from "@vanilla-extract/css";
+
+export const noticeStyle = style({
+  selectors: {
+    "&:not(:first-child)": {
+      marginTop: "24px",
+    },
+  },
+});
+
+export const labelStyle = style({
+  marginTop: "36px",
+});
